CRNS-107: decoupled action sheet from message input
