+++ src/components/MessageInput/index.js
@@ -1,3 +1,4 @@
+export { default as ActionSheet } from './ActionSheet';
 export { default as AttachButton } from './AttachButton';
 export { default as FileUploadPreview } from './FileUploadPreview';
 export { default as ImageUploadPreview } from './ImageUploadPreview';
