+++ styleguide.config.js
@@ -108,6 +108,7 @@ module.exports = {
         'src/components/MessageInput/ImageUploadPreview.js',
         'src/components/MessageInput/FileUploadPreview.js',
         'src/components/MessageInput/SendButton.js',
+        'src/components/MessageInput/ActionSheet.js',
         'src/components/MessageInput/AttachButton.js',
         'src/components/AutoCompleteInput/AutoCompleteInput.js',
         'src/components/AutoCompleteInput/CommandsItem.js',
